Project Euler+: Problem #27
Problem Statement

This problem is a programming version of Problem 27 from
projecteuler.net

Euler published the remarkable quadratic formula:
n2+n+41

It turns out that the formula will produce 40 primes for the
consecutive values n=0 to 39. However, when n=40, 402+40+41=40(40+1)+41
is divisible by 41, and certainly when n=41, 412+41+41 is clearly
divisible by 41.

Using computers, the incredible formula n2−79n+1601 was discovered,
which produces 80 primes for the consecutive values n=0 to 79. The
product of the coefficients, −79 and 1601, is −126479.

Considering quadratics of the form:

n2+an+b, where |a|≤N and |b|≤N

where |n| is the modulus/absolute value of n
e.g. |11|=11 and |−4|=4

Find the coefficients, a and b, for the quadratic expression that
produces the maximum number of primes for consecutive values of n,
starting with n=0.

Note For this challenge you can assume solution to be unique.

Input Format

The first line contains an integer N.

Output Format

Print the value of a and b separated by space.

Constraints
42≤N≤2000

Sample Input

42

Sample Output

-1 41

Explanation

for a=−1 and b=41, you get 42 primes.