Project Euler+: Problem #28
Problem Statement

This problem is a programming version of Problem 28 from
projecteuler.net

Starting with the number 1 and moving to the right in a clockwise
direction a 5 by 5 spiral is formed as follows:

21201918172276516238141524923142510111213

It can be verified that the sum of the numbers on the diagonals is 101.
What is the sum of the numbers on the diagonals in a N×N, (N is odd)
spiral formed in the same way?
As the sum will be huge you have to print the result mod (109+7)

Input Format
The first line contains an integer T , i.e., number of test cases.
Next T lines will contain an integer N.

Output Format
Print the values corresponding to each test case.

Constraints
1≤T≤105
1≤N<1018,N is odd

Sample Input

2
3
5

Sample Output

25
101