Project Euler+: Problem #30
Problem Statement

This problem is a programming version of Problem 30 from
projecteuler.net

Surprisingly there are only three numbers that can be written as the
sum of fourth powers of their digits:

1634=14+64+34+448208=84+24+04+849474=94+44+74+44

As 1 = 1^4 is not a sum it is not included.
The sum of these numbers is 1634+8208+9474=19316.

Find the sum of all the numbers that can be written as the sum of Nth
powers of their digits.

Input Format
Input contains an integer N

Output Format
Print the answer corresponding to the test case.

Constraints
3≤N≤6

Sample Input

4

Sample Output

19316