Project Euler+: Problem #43
Problem Statement

This problem is a programming version of Problem 43 from
projecteuler.net

The number, 1406357289, is a 0 to 9 pandigital number because it is
made up of each of the digits 0 to 9 in some order, but it also has a
rather interesting sub-string divisibility property.

Let d1 be the 1st digit, d2 be the 2nd digit, and so on. In this way,
we note the following:
d2d3d4 is divisible by 2d3d4d5 is divisible by 3d4d5d6 is divisible by
5d5d6d7 is divisible by 7d6d7d8 is divisible by 11d7d8d9 is divisible
by 13d8d9d10 is divisible by 17

Find the sum of all 0 to N pandigital numbers with this property.

Input Format
Input contains an integer N

Output Format
Print the answer corresponding to the test case.

Constraints
3≤N≤9

Sample Input

3

Sample Output

22212