Project Euler+: Problem #46
Problem Statement

This problem is a programming version of Problem 46 from
projecteuler.net

It was proposed by Christian Goldbach that every odd composite number
can be written as the sum of a prime and twice a square.
91521252733=7+2×12=7+2×22=3+2×32=7+2×32=19+2×22=31+2×12

It turns out that the conjecture was false as you'll discover some
values can't be represented as a sum of prime and twice a square.
You are given N, print the number of ways N can be represented as a sum
of prime and twice a square.
Example 15 can be represented in two ways as 15=7+2×22 and 15=13+2×12

Input Format
The first line contains an integer T , i.e., number of test cases.
Next T lines will contain an integer N.

Output Format
Print the values corresponding to each test case.

Constraints
1≤T≤100
9≤N<5×105
N∈{odd composite number}

Sample Input

2
9
15

Sample Output

1
2