Project Euler+: Problem #77
Problem Statement

This problem is a programming version of Problem 77 from
projecteuler.net

It is possible to write ten as the sum of primes in exactly five
different ways:
7+35+55+3+23+3+2+22+2+2+2+2

You are given N, in how many ways can N be written as sum of 1 or more
primes?

Input Format
First line of the input contains T, which is number of testcases.
Each testcase contains N.

Constraints
1≤T≤100
2≤N≤1000

Output Format
Print the output corresponding to each testcase on a new line.

Sample Input

2
5
10

Sample Output

2
5