Project Euler+: Problem #125
Problem Statement

The palindromic number 595 is interesting because it can be written as
the sum of consecutive squares: 62+72+82+92+102+112+122 .

The palindromic number 696 is also nice because it can be written as
102+122+142+162, where the bases form an arithmetic progression with
common difference 2.

There are exactly eleven palindromes below one-thousand that can be
written as consecutive square sums, and the sum of these palindromes is
4164. Note that 1=02+12 has not been included as this problem is
concerned with the squares of positive integers. Also, there has to be
at least two terms in the sum.

Given N and d, find the sum of all the numbers less than N that are
both palindromic and can be written as the sum of squares whose bases
form an arithmetic progression with common difference d.

Input Format

The first line of input contains T, the number of test cases.

Each test case consists of a single line containing two integers N and
d, separated by a space.

Constraints

1≤T≤20
1≤N≤109
1≤d≤109

Output Format

For each test case, output a single line containing a single integer,
the answer for that test case.

Sample Input

2
1000 1
1000 2

Sample Output

4164
3795

Explanation

The first test case corresponds to the example given in the problem
statement.

In the second test case, d=2, and there are 6 such numbers less than
1000. Two such numbers are:

696=102+122+142+162
969=12+32+52+72+92+112+132+152+172