Project Euler+: Problem #74
This problem is a programming version of Problem 74 from
projecteuler.net

The number 145145 is well known for the property that the sum of the
factorial of its digits is equal to 145145:

1!+4!+5!=1+24+120=145
1!+4!+5!=1+24+120=145

Perhaps less well known is 169, in that it produces the longest chain
of numbers that link back to 169; it turns out that there are only
three such loops that exist:

169→363601→1454→169871→45361→871872→45362→872
169→363601→1454→169871→45361→871872→45362→872

It is not difficult to prove that EVERY starting number will eventually
get stuck in a loop. For example,

69→363600→1454→169→363601(→1454)78→45360→871→45361(→871)540→145(→145)
69→363600→1454→169→363601(→1454)78→45360→871→45361(→871)540→145(→145)

Starting with 6969 produces a chain of five non-repeating terms, but
the longest non-repeating chain with a starting number below one
million is sixty terms.

For a given length LL and limit NN print all the integers ≤N≤N which
have chain length LL

Input Format
First line contains TT, followed by TT lines.
Each line contains NN and LL separated by space.

Constraints
1≤T≤101≤T≤10
10≤N≤100000010≤N≤1000000
1≤L≤601≤L≤60

Output Format
Print the integers separated by space for each testcase. Where there
are no such number for a given LL, print -1.

Sample Input

10
221 7
147 1
258 4
265 8
210 2
175 7
29 2
24 3
273 4
261 4

Sample Output

24 42 104 114 140 141
1 2 145
78 87 196 236
4 27 39 72 93 107 117 170 171
0 10 11 154
24 42 104 114 140 141
0 10 11
-1
78 87 196 236 263
78 87 196 236